datapulse-analytics/requirements.txt
# ─── Core ───────────────────────────────────────────────────────────────────
pandas==2.1.4
numpy==1.26.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0

# ─── Data Generation ────────────────────────────────────────────────────────
faker==21.0.0

# ─── Data Quality ───────────────────────────────────────────────────────────
great-expectations==0.18.8

# ─── dbt ────────────────────────────────────────────────────────────────────
dbt-core==1.7.3
dbt-postgres==1.7.3

# ─── Dashboard ──────────────────────────────────────────────────────────────
streamlit==1.29.0
plotly==5.17.0
altair==5.2.0

# ─── Airflow (install separately — heavy dependency) ────────────────────────
# apache-airflow==2.7.3

# ─── Testing ────────────────────────────────────────────────────────────────
pytest==7.4.3
pytest-cov==4.1.0

# ─── Dev Tools ──────────────────────────────────────────────────────────────
black==23.11.0
flake8==6.1.0
isort==5.12.0
